+++ test/qwen_test.exs
@@ -1,6 +1,9 @@
 defmodule QwenTest do
   use ExUnit.Case
-  doctest Qwen
+  import Qwen.Sigils
+
+  # here we only specified some of the modules use doctest to AVOID the call of REST API
+  doctest Qwen.Sigils
 
   test "greets the world" do
     assert :world == :world